+++ src/components/Space.tsx
@@ -1,7 +1,7 @@
 // components/Space.tsx
 import React from 'react';
-import { Text, View, TouchableOpacity } from 'react-native';
-import { spaceStyle } from './Styles';
+import {Text, View, TouchableOpacity} from 'react-native';
+import {spaceStyle} from '../improve/Styles';
 
 interface SpaceProps {
   spaceID: number;
@@ -9,14 +9,13 @@ interface SpaceProps {
   setTurn: (spaceID: number) => void;
 }
 
-const Space: React.FC<SpaceProps> = ({ spaceID, spaceValue, setTurn }) => {
+const Space: React.FC<SpaceProps> = ({spaceID, spaceValue, setTurn}) => {
   return (
     <View style={spaceStyle.container}>
       <TouchableOpacity
         style={spaceStyle.space}
         onPress={() => setTurn(spaceID)}
-        disabled={!!spaceValue}
-      >
+        disabled={!!spaceValue}>
         <Text style={spaceStyle.spaceValue}>{spaceValue}</Text>
       </TouchableOpacity>
     </View>
