+++ src/interfaces/Board.ts
@@ -1,0 +1,4 @@
+export interface BoardProps {
+  turns: Record<number, string>;
+  onTurn: (index: number) => void;
+}
